# Path as array of props instead of dot-notation string