fixes #44