@DarrenPaulWright 2019-09-13
Try running the benchmarks before and after this change. Pay attention particularly to nested read, object write, and array write in the first two suites where they are taking two to three times as long to run. 

The problem is Array.concat. Try the following and re-run the benchmarks: 
```js
const concatPath = (path = [], prop) => {
	if (prop === null || prop === undefined) {
		return path;
	}

	const output = [...path];
	output.push(prop);

	return output;
};
```

I still want to keep the previous string path as the default behavior, so please add an option to opt-in to getting the path as an array.