@qwelias 2019-09-14
Agree with concern regarding `concat`, fixed. But interestingly on node12:
- with `Array#push`: `npm run bench  100.03s user 1.71s system 94% cpu 1:47.25 total`
- with `Array#concat`: `npm run bench  99.01s user 1.59s system 94% cpu 1:46.31 total`
- original: `npm run bench  95.53s user 1.73s system 94% cpu 1:43.16 total`

I guess V8 optimizations really kick in.

Not making array path optional tho, as I think it's inherently wrong.
`walkPath` will bug out if:
- props have `.`
- `ignoreSymbols == false` and path contains a symbol