@qwelias 2019-09-14
Nvm my notes regarding performance, `matcha` is broken on node12.