@SReject 2019-11-17
> Not making array path optional tho, as I think it's inherently wrong.

While I agree with the statement in general, this project was already published to npm with the feature of path being a string.

So either the main version number in package.json needs bumped or you can alter the handleChange function to

```js
if (!inApply) {

    // if path should be represented as array
    if (options.pathAsArray) {
        path = concatPath(path, property);

    // otherwise convert the path array to a dot-notated string; escape dots in property names
    } else {
        let pathAsString = '';
        path.forEach(property => {
            if (property.toString) {
                if (pathAsString) {
                    pathAsString += '.';
                }
                pathAsString += property.toString().replace(/\./g, "\\.");
            }
        });

        if (property.toString) {
            if (pathAsString) {
                pathAsString += '.';
            }
            pathAsString += property.toString().replace(/\./g, "\\.");
        }

        path = pathAsString;
    }

    onChange(path, value, previous);
    return;
}
```