@qwelias 2019-11-17
@SReject Not opposed to major bump btw.
Also feel free to branch off of my changes and submit another PR if you want to build on top of it :)