@SReject 2019-11-17
@qwelias I'm not against a major bump either. Just figured I'd spotlight that side of the issue in case it was overlooked.