@DarrenPaulWright 2020-03-12
An option "pathAsArray" is being added in #52